+++ virtualizarr/readers/tiff.py
@@ -1,4 +1,5 @@
 import warnings
+from pathlib import Path
 from typing import Iterable, Mapping, Optional
 
 from xarray import Dataset, Index
@@ -57,6 +58,7 @@ class TIFFVirtualBackend(VirtualBackend):
             refs,
             loadable_variables,
             drop_variables,
+            fs_root=Path.cwd().as_uri(),
         )
 
         loadable_vars, indexes = open_loadable_vars_and_indexes(