+++ virtualizarr/tests/test_readers/test_dmrpp.py
@@ -1,3 +1,4 @@
+import os
 import textwrap
 from pathlib import Path
 from xml.etree import ElementTree as ET
@@ -19,6 +20,7 @@ urls = [
     )
     # TODO: later add MUR, SWOT, TEMPO and others by using kerchunk JSON to read refs (rather than reading the whole netcdf file)
 ]
+
 
 DMRPP_XML_STRINGS = {
     "basic": textwrap.dedent(
@@ -390,3 +392,53 @@ def test_parse_chunks(
         lengths=expected_lengths, offsets=expected_offsets, paths=expected_paths
     )
     assert result == expected
+
+
+@pytest.fixture
+def basic_dmrpp_temp_filepath(tmp_path: Path) -> Path:
+    # TODO generalize key here? Would require the factory pattern
+    # (https://docs.pytest.org/en/stable/how-to/fixtures.html#factories-as-fixtures)
+    drmpp_xml_str = DMRPP_XML_STRINGS["basic"]
+
+    # TODO generalize filename here?
+    filepath = tmp_path / "test.nc.dmrpp"
+
+    with open(filepath, "w") as f:
+        f.write(drmpp_xml_str)
+
+    return filepath
+
+
+class TestRelativePaths:
+    def test_absolute_path_to_dmrpp_file_containing_relative_path(
+        self,
+        basic_dmrpp_temp_filepath: Path,
+    ):
+        vds = open_virtual_dataset(
+            str(basic_dmrpp_temp_filepath), indexes={}, filetype="dmrpp"
+        )
+        path = vds["x"].data.manifest["0"]["path"]
+
+        # by convention, if dmrpp file path is {PATH}.nc.dmrpp, the data filepath should be {PATH}.nc
+        # and the manifest should only contain absolute file URIs
+        expected_datafile_path_uri = basic_dmrpp_temp_filepath.as_uri().removesuffix(
+            ".dmrpp"
+        )
+        assert path == expected_datafile_path_uri
+
+    def test_relative_path_to_dmrpp_file(self, basic_dmrpp_temp_filepath: Path):
+        # test that if a user supplies a relative path to a DMR++ file we still get an absolute path in the manifest
+        relative_dmrpp_filepath = os.path.relpath(
+            str(basic_dmrpp_temp_filepath), start=os.getcwd()
+        )
+
+        vds = open_virtual_dataset(
+            relative_dmrpp_filepath, indexes={}, filetype="dmrpp"
+        )
+        path = vds["x"].data.manifest["0"]["path"]
+
+        # by convention, if dmrpp file path is {PATH}.nc.dmrpp, the data filepath should be {PATH}.nc
+        expected_datafile_path_uri = basic_dmrpp_temp_filepath.as_uri().removesuffix(
+            ".dmrpp"
+        )
+        assert path == expected_datafile_path_uri